migrated servo_board_config.h into components and respective individual component config files
