+++ Firmware/new-member-projects/Justin_F/f401re-receiver/.settings/stm32cubeide.project.prefs
@@ -1,5 +1,5 @@
 635E684B79701B039C64EA45C3F84D30=4EFB57242C66A2F320AB8CC1075F7759
 66BE74F758C12D739921AEA421D593D3=0
-8DF89ED150041C4CBC7CB9A9CAA90856=02E468D49D9197516C55386B2F5F5AD4
-DC22A860405A8BF2F2C095E5B6529F12=02E468D49D9197516C55386B2F5F5AD4
+8DF89ED150041C4CBC7CB9A9CAA90856=E8737F6BA2D4416BB504CBFF6ED8C87B
+DC22A860405A8BF2F2C095E5B6529F12=562D395E1C0EB8DB5204A8AAF72461D6
 eclipse.preferences.version=1
